extent the class by using interface
it is possible by using extensibility which means instead of changing the code in the existing classes simply add in new classes that change the behavior of the system. for example,  here we define the logger interface. whenever we add new behavior,  we just need to add new classes, ConsoleLogger, FileLogger, and DbLogger.